data/repos/hocr_(math_capture)/content/pyproject.toml
[project]
name = "hocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "mcp>=1.6.0",
    "openai>=1.76.2",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
]
